+++ apps/lib/components/FormBuilder/fields/index.ts
@@ -8,14 +8,14 @@ export {
 } from "./TextField";
 export { SelectField, MultiSelectField } from "./SelectField";
 export {
-  SwitchField,
   CheckboxField,
-  RadioField,
   SegmentedField,
   ToggleGroupField,
   RadioCardsField,
   ToggleButtonField,
 } from "./ChoiceFields";
+export { RadioListField, CheckboxGroupField } from "./OptionListFields";
+export { SwitchBoxField } from "./SwitchBoxField";
 export { DateField } from "./DateField";
 export { CalendarField } from "./CalendarField";
 export { OtpField } from "./OtpField";